auth0: Implementing first version of call for auth0
The SDK is supposed to provide methods that can be called to be as
simple to use as the Auth0 API we are wrapping. This adds the methods
CRUD for users and review how the call are made to have something
as generic as possible.

Currently that will work with almost all Auth0 User's API methods,
but if we want to implement all of their methods, we might have
to slightly modify it, as they sometimes include extra 'directory'
such as for:
* Delete a user's multifactor provider
* Unlink a user account
* or Link a user account

The methods for CRUD on users includes basic checks of nil/empty.
Most of the code (94.1%) is also tested. We should be added as a hook
ideally as well as go vet to have automatic check/coverage.