travis: Adding a gofmt check on the Makefile for the all rule
The 'gofmt' and 'go fmt' binaries are returning 0 even when they
modify/only check files. Adding this verification on the Makefile allows
the CI to validate or not that the formatting of the code is respected -
or not.

Modifying the .travis.yml to add check of formatting via the Makefile